requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
werkzeug
google-generativeai
